Fork an integration branch to record what we actually distribute
Initial changes to kick that off:
Don't automatically configure the tree when autogen.sh is run.
Don't leave behind the autom4te.cache or copies of m4 added to aclocal either.
Don't ignore the things that we do want to include in this branch (but not in
the master source tree).  But do ignore vim's inner machinations.
Don't export our ignorance.