Clean up after dh_installudev fail
Closes: #565126, at least for anyone who didn't modify rules themselves.
People who did will need to edit and keep (some of) that file anyhow.

Since the version of that file they might have could be one of many,
depending on which version of the package they had updated from to
one where it had moved, we need to query dpkg/status for the expected
hash of an unmodified rules file on their system.